@@ -24,70 +24,70 @@
   "dependencies": {
     "classnames": "2.2.5",
     "clipboard": "1.7.1",
-    "codemirror": "5.32.0",
+    "codemirror": "5.33.0",
     "color": "2.0.1",
-    "core-js": "2.5.1",
+    "core-js": "2.5.3",
     "crypto-js": "3.1.9-1",
     "file-saver": "1.3.3",
     "golden-layout": "1.5.9",
-    "inversify": "4.5.1",
-    "inversify-inject-decorators": "3.0.2",
+    "inversify": "4.9.0",
+    "inversify-inject-decorators": "3.1.0",
     "isomorphic-fetch": "2.2.1",
     "jed": "1.1.1",
     "jquery": "3.2.1",
     "keymaster": "1.6.2",
     "markdown-it": "8.4.0",
-    "moment": "2.19.2",
+    "moment": "2.20.1",
     "po2json": "0.4.5",
-    "react": "16.1.1",
-    "react-dom": "16.1.1",
+    "react": "16.2.0",
+    "react-dom": "16.2.0",
     "react-router": "4.2.0",
     "react-router-dom": "4.2.2",
     "reflect-metadata": "0.1.10",
-    "rxjs": "5.5.2",
+    "rxjs": "5.5.6",
     "spark-md5": "3.0.0",
-    "tslib": "1.8.0",
+    "tslib": "1.8.1",
     "tv4": "1.3.0"
   },
   "devDependencies": {
-    "@types/chai": "4.0.5",
+    "@types/chai": "4.1.0",
     "@types/chai-as-promised": "7.1.0",
     "@types/classnames": "2.2.3",
     "@types/clipboard": "1.5.35",
-    "@types/codemirror": "0.0.51",
+    "@types/codemirror": "0.0.55",
     "@types/color": "2.0.0",
-    "@types/core-js": "0.9.43",
-    "@types/crypto-js": "3.1.37",
-    "@types/enzyme": "3.1.5",
+    "@types/core-js": "0.9.44",
+    "@types/crypto-js": "3.1.38",
+    "@types/enzyme": "3.1.6",
     "@types/enzyme-adapter-react-16": "1.0.1",
     "@types/esprima": "4.0.1",
     "@types/file-saver": "1.3.0",
-    "@types/glob": "5.0.33",
+    "@types/glob": "5.0.34",
     "@types/history": "4.6.2",
     "@types/isomorphic-fetch": "0.0.34",
-    "@types/jquery": "3.2.16",
+    "@types/jquery": "3.2.18",
     "@types/keymaster": "1.6.28",
     "@types/markdown-it": "0.0.4",
-    "@types/mocha": "2.2.44",
-    "@types/node": "8.0.53",
-    "@types/react": "16.0.25",
+    "@types/mocha": "2.2.46",
+    "@types/node": "9.3.0",
+    "@types/react": "16.0.34",
     "@types/react-dom": "16.0.3",
-    "@types/react-router": "4.0.18",
-    "@types/react-router-dom": "4.2.2",
+    "@types/react-router": "4.0.20",
+    "@types/react-router-dom": "4.2.3",
     "@types/react-test-renderer": "16.0.0",
     "@types/tv4": "1.2.28",
-    "archiver": "2.1.0",
-    "awesome-typescript-loader": "3.4.0",
+    "archiver": "2.1.1",
+    "awesome-typescript-loader": "3.4.1",
     "chai": "4.1.2",
     "chai-as-promised": "7.1.1",
     "clean-webpack-plugin": "0.1.17",
-    "copy-webpack-plugin": "4.2.1",
-    "css-loader": "0.28.7",
-    "enzyme": "3.2.0",
-    "enzyme-adapter-react-16": "1.1.0",
+    "copy-webpack-plugin": "4.3.1",
+    "css-loader": "0.28.8",
+    "enzyme": "3.3.0",
+    "enzyme-adapter-react-16": "1.1.1",
     "esprima": "4.0.0",
     "extract-text-webpack-plugin": "3.0.2",
-    "file-loader": "1.1.5",
+    "file-loader": "1.1.6",
     "git-revision-webpack-plugin": "2.5.1",
     "glob": "7.1.2",
     "handlebars": "4.0.11",
@@ -98,37 +98,37 @@
     "istanbul-instrumenter-loader": "3.0.0",
     "istanbul-lib-instrument": "1.9.1",
     "json-loader": "0.5.7",
-    "karma": "1.7.1",
+    "karma": "2.0.0",
     "karma-bamboo-reporter": "0.1.2",
     "karma-chrome-launcher": "2.2.0",
     "karma-coverage": "1.1.1",
-    "karma-firefox-launcher": "1.0.1",
+    "karma-firefox-launcher": "1.1.0",
     "karma-mocha": "1.3.0",
     "karma-mocha-reporter": "2.2.5",
     "karma-sourcemap-loader": "0.3.7",
-    "karma-webpack": "2.0.6",
-    "lint-staged": "5.0.0",
-    "mocha": "4.0.1",
-    "node-sass": "4.7.2",
-    "pofile": "1.0.9",
-    "prettier": "1.8.2",
+    "karma-webpack": "2.0.9",
+    "less": "2.7.3",
+    "less-loader": "4.0.5",
+    "lint-staged": "6.0.0",
+    "mocha": "4.1.0",
+    "pofile": "1.0.10",
+    "prettier": "1.10.2",
     "raw-loader": "0.5.1",
-    "react-test-renderer": "16.1.1",
-    "remap-istanbul": "0.9.5",
-    "sass-loader": "6.0.6",
-    "style-loader": "0.19.0",
-    "tslint": "5.8.0",
+    "react-test-renderer": "16.2.0",
+    "remap-istanbul": "0.10.0",
+    "style-loader": "0.19.1",
+    "tslint": "5.9.1",
     "tslint-loader": "3.5.3",
-    "typescript": "2.6.1",
+    "typescript": "2.6.2",
     "url-loader": "0.6.2",
-    "webpack": "3.8.1",
-    "webpack-bundle-analyzer": "2.9.1",
-    "webpack-dev-server": "2.9.4",
+    "webpack": "3.10.0",
+    "webpack-bundle-analyzer": "2.9.2",
+    "webpack-dev-server": "2.11.0",
     "webpack-notifier": "1.5.0"
   },
   "lint-staged": {
     "linters": {
-      "*.{ts,tsx,js,json,css,scss}": ["prettier --write", "git add"]
+      "*.{ts,tsx,js,json,css,less}": ["prettier --write", "git add"]
     }
   }
 }
