@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as cn from "classnames";
 
-import "./no-data.component.scss";
+import "./no-data.component.less";
 
 interface IComponentProps {
   title: JSX.Element | string;
