@@ -17,7 +17,7 @@ import { FileSelectorComponent } from "app/ui/_generic/file-selector.component";
 import { resolveInject } from "app/di";
 import { Routes } from "app/routes";
 import { UserInfo } from "app/services/login/user-info";
-import { TurtlesService } from "app/services/customizations/turtles.service";
+import { TurtlesService, TurtleInfo, TurtleSize } from "app/services/customizations/turtles.service";
 import { Theme, ThemesService } from "app/services/customizations/themes.service";
 import { ProgramsExportImportService } from "app/services/gallery/programs-export-import.service";
 import { LocalizationService, ILocaleInfo, _T } from "app/services/customizations/localization.service";
@@ -29,6 +29,13 @@ import { INotificationService } from "app/services/infrastructure/notification.s
 import { IUserLibraryRepository } from "app/services/gallery/personal-gallery-localstorage.repository";
 import { ProgramsHtmlSerializerService } from "app/services/gallery/programs-html-serializer.service";
 import { IEventsTrackingService, EventAction } from "app/services/infrastructure/events-tracking.service";
+import { ensure } from "app/utils/syntax-helpers";
+import { SimpleSelectComponent } from "app/ui/_generic/simple-select.component";
+
+class LocaleSelector extends SimpleSelectComponent<ILocaleInfo> {}
+class ThemeSelector extends SimpleSelectComponent<Theme> {}
+class TurtleSelector extends SimpleSelectComponent<TurtleInfo> {}
+class TurtleSizeSelector extends SimpleSelectComponent<TurtleSize> {}
 
 interface IComponentState {
   userInfo: UserInfo;
@@ -47,9 +54,23 @@ const codeSamples = [
   "pu setxy -40 -20 pd repeat 8 [fd 40 rt 360/8]",
   "repeat 10 [repeat 8 [fd 20 rt 360/8] rt 360/10]",
   "repeat 14 [fd repcount*8 rt 90]",
-  "window repeat 10 [fd 5 * repcount repeat 3 [fd 18 rt 360/3] rt 360/10]",
-  "pu setxy -20 -20 pd repeat 8 [rt 45 repeat 4 [repeat 90 [fd 1 rt 2] rt 90]]",
-  "repeat 10 [fd 10 rt 90 fd 10 lt 90]"
+  "repeat 10 [fd 5 * repcount repeat 3 [fd 18 rt 360/3] rt 360/10]",
+  "repeat 10 [fd 10 rt 90 fd 10 lt 90]",
+  `
+penup
+setxy -80 0
+repeat 10 [
+  arc 360 20
+  fd 40
+  rt 36
+]
+  `,
+  `
+  rt 18
+repeat 5 [
+	fd 100
+  	rt 144
+]`
 ];
 
 export class UserProfilePageComponent extends React.Component<IComponentProps, IComponentState> {
@@ -100,7 +121,7 @@ export class UserProfilePageComponent extends React.Component<IComponentProps, I
       callActionSafe(this.errorHandler, async () => this.userSettingsService.get())
     ]);
     if (programs && userSettings) {
-      const theme = this.themeService.getTheme(userSettings.themeName);
+      const theme = this.themeService.getActiveTheme() || this.themeService.getTheme("Default");
       const locale = this.localizationService.getLocaleById(userSettings.localeId);
       const turtleImage = this.turtleCustomizationService.getTurtleImage(userSettings.turtleId);
 
@@ -119,7 +140,11 @@ export class UserProfilePageComponent extends React.Component<IComponentProps, I
       this.exportInportService.exportAll(this.programsReporitory)
     );
     if (programs) {
-      const html = new ProgramsHtmlSerializerService().serialize(programs);
+      const html = await new ProgramsHtmlSerializerService().serialize(
+        programs,
+        this.currentUser.getLoginStatus().userInfo.attributes.name,
+        this.currentUser.getLoginStatus().userInfo.attributes.imageUrl
+      );
       const blob = new Blob([html], { type: "text/plain;charset=utf-8" });
       FileSaver.saveAs(blob, `my-logo-programs.html`);
     }
@@ -145,170 +170,184 @@ export class UserProfilePageComponent extends React.Component<IComponentProps, I
 
   render(): JSX.Element {
     return (
-      <div>
+      <div className="ex-page-container">
         <MainMenuComponent />
-        <div className="container">
-          <br />
-          <PageHeaderComponent title={_T("User settings")} />
+        <div className="ex-page-content">
+          <div className="container">
+            <br />
+            <PageHeaderComponent title={_T("User settings")} />
 
-          {this.state.userSettings &&
-            this.state.currentLocale &&
-            this.state.theme &&
-            this.state.userInfo && (
-              <div className="tile is-ancestor">
-                <div className="tile is-6 is-vertical is-parent">
-                  <div className="tile is-child box">
-                    <p>
-                      <strong>{_T("Name")}:</strong> {this.state.userInfo.attributes.name}
-                    </p>
-                    <p>
-                      <strong>{_T("Email")}:</strong> {this.state.userInfo.attributes.email}
-                    </p>
+            {this.state.userSettings &&
+              this.state.currentLocale &&
+              this.state.theme &&
+              this.state.userInfo && (
+                <div className="columns">
+                  <div className="column">
+                    <div className="card">
+                      <div className="card-content">
+                        <div className="media">
+                          {this.state.userInfo.attributes.imageUrl && (
+                            <div className="media-left">
+                              <figure className="image is-48x48">
+                                <img src={this.state.userInfo.attributes.imageUrl} alt="User image" />
+                              </figure>
+                            </div>
+                          )}
+
+                          <div className="media-content">
+                            <p className="title is-4">{this.state.userInfo.attributes.name || _T("Guest")}</p>
+                            <p className="subtitle is-6">{this.state.userInfo.attributes.email}</p>
+                          </div>
+                        </div>
+
+                        <div className="content">
+                          {this.currentUser.getLoginStatus().isLoggedIn ? (
+                            <>
+                              <p>
+                                {_T("You are logged in with %s authentication.", {
+                                  value: this.state.userInfo.attributes.authProvider.toLowerCase()
+                                })}
+                              </p>
+                            </>
+                          ) : (
+                            <>
+                              <p>{_T("NOT_LOGGED_IN_TEXT_BLOCK")}</p>
+                              <Link to={Routes.loginRoot.build({})} className="button is-primary">
+                                {_T("Log in")}
+                              </Link>
+                            </>
+                          )}
+                        </div>
+
+                        <br />
+
+                        <div className="field">
+                          <label className="label" htmlFor="language-selector">
+                            {_T("Language")}
+                          </label>
+                          <div className="control">
+                            <div className="select">
+                              <LocaleSelector
+                                items={this.localizationService.getSupportedLocales()}
+                                selectedItem={this.state.currentLocale}
+                                getItemIdentifier={x => x.id}
+                                renderItem={x => x.name}
+                                idAttr="language-selector"
+                                selectionChanged={selectedLocation => {
+                                  if (selectedLocation) {
+                                    setTimeout(async () => {
+                                      await this.userSettingsService.update({ localeId: selectedLocation.id });
+                                      // refresh browser window
+                                      window.location.reload(true);
+                                    }, 0);
+                                  }
+                                }}
+                              />
+                            </div>
+                          </div>
+                        </div>
+
+                        <br />
+
+                        <div className="field">
+                          <label className="label" htmlFor="theme-selector">
+                            {_T("User interface theme")}
+                          </label>
+                          <div className="control">
+                            <div className="select">
+                              <ThemeSelector
+                                items={this.themeService.getAllThemes()}
+                                selectedItem={this.state.theme}
+                                getItemIdentifier={x => x.name}
+                                renderItem={x => `${x.name} - ${x.description}`}
+                                idAttr="theme-selector"
+                                selectionChanged={async selectedTheme => {
+                                  if (selectedTheme) {
+                                    await this.userSettingsService.update({ themeName: selectedTheme.name });
+                                    this.themeService.setActiveTheme(selectedTheme);
+                                    await this.loadData();
+                                    this.setRandomCode();
+                                  }
+                                }}
+                              />
+                            </div>
+                          </div>
+                        </div>
+
+                        <br />
+
+                        <label className="label">{_T("Turtle")}</label>
+                        <div className="field is-grouped">
+                          <div className="control">
+                            <div className="select">
+                              <TurtleSelector
+                                items={this.turtleCustomizationService.getAllTurtles()}
+                                selectedItem={this.turtleCustomizationService
+                                  .getAllTurtles()
+                                  .find(x => x.id === ensure(this.state.userSettings).turtleId)}
+                                getItemIdentifier={x => x.id}
+                                renderItem={x => x.getName()}
+                                selectionChanged={async newTurtle => {
+                                  if (newTurtle) {
+                                    await this.userSettingsService.update({ turtleId: newTurtle.id });
+                                    await this.loadData();
+                                    this.setRandomCode();
+                                  }
+                                }}
+                              />
+                            </div>
+                          </div>
+                          <div className="control">
+                            <div className="select">
+                              <TurtleSizeSelector
+                                items={this.turtleCustomizationService.getTurtleSizes()}
+                                selectedItem={this.turtleCustomizationService
+                                  .getTurtleSizes()
+                                  .find(x => x.size === ensure(this.state.userSettings).turtleSize)}
+                                getItemIdentifier={x => x.size.toString()}
+                                renderItem={x => x.description}
+                                selectionChanged={async newSize => {
+                                  if (newSize) {
+                                    await this.userSettingsService.update({ turtleSize: newSize.size });
+                                    await this.loadData();
+                                    this.setRandomCode();
+                                  }
+                                }}
+                              />
+                            </div>
+                          </div>
+                        </div>
+
+                        <br />
+
+                        <label className="label">{_T("Personal library")}</label>
+                        <p className="help">
+                          <span>
+                            {_T("You have %d program", {
+                              plural: "You have %d programs",
+                              value: this.state.programCount
+                            })}
+                          </span>
+                        </p>
+                        <div className="field is-grouped is-grouped-multiline">
+                          <p className="control">
+                            <a className="button" onClick={this.doExport}>
+                              {_T("Export")}
+                            </a>
+                          </p>
+                          <p className="control">
+                            <FileSelectorComponent
+                              className={cn({ "is-loading": this.state.isImportingInProgress })}
+                              buttonText={_T("Import")}
+                              onFileTextContentReady={this.onImport}
+                            />
+                          </p>
+                        </div>
+                      </div>
+                    </div>
                   </div>
-                  <div className="tile is-child box">
-                    <div className="field">
-                      <label className="label">{_T("Language")}</label>
-                      <div className="control">
-                        <div className="select">
-                          <select
-                            id="localeselector"
-                            value={this.state.currentLocale.id}
-                            onChange={event => {
-                              const selectedLocation = this.localizationService
-                                .getSupportedLocales()
-                                .find(loc => loc.id === event.target.value);
-                              if (selectedLocation) {
-                                setTimeout(async () => {
-                                  await this.userSettingsService.update({ localeId: selectedLocation.id });
-                                  // refresh browser window
-                                  window.location.reload(true);
-                                }, 0);
-                              }
-                              return {};
-                            }}
-                          >
-                            {this.localizationService.getSupportedLocales().map(loc => {
-                              return (
-                                <option key={loc.id} value={loc.id}>
-                                  {loc.name}
-                                </option>
-                              );
-                            })}
-                          </select>
-                        </div>
-                      </div>
-                    </div>
 
-                    <div className="field">
-                      <label className="label">{_T("User interface theme")}</label>
-                      <div className="control">
-                        <div className="select">
-                          <select
-                            id="themeselector"
-                            value={this.state.theme.name}
-                            onChange={event => {
-                              const selectedTheme = this.themeService
-                                .getAllThemes()
-                                .find(t => t.name === event.target.value);
-                              if (selectedTheme) {
-                                setTimeout(async () => {
-                                  await this.userSettingsService.update({ themeName: selectedTheme.name });
-                                  window.localStorage.setItem(
-                                    (window as any).appThemeNameLocalStorageKey,
-                                    JSON.stringify(selectedTheme)
-                                  );
-                                  // refresh browser window
-                                  window.location.reload(true);
-                                }, 0);
-                              }
-                              return {};
-                            }}
-                          >
-                            {this.themeService.getAllThemes().map(t => {
-                              return (
-                                <option key={t.name} value={t.name}>
-                                  {t.name} - {t.description}
-                                </option>
-                              );
-                            })}
-                          </select>
-                        </div>
-                      </div>
-                    </div>
-
-                    <label className="label">{_T("Turtle")}</label>
-                    <div className="field is-grouped">
-                      <div className="control">
-                        <div className="select">
-                          <select
-                            id="turtleSelector"
-                            value={this.state.userSettings.turtleId}
-                            onChange={async event => {
-                              const value = event.target.value;
-                              await this.userSettingsService.update({ turtleId: value });
-                              await this.loadData();
-                              this.setRandomCode();
-                            }}
-                          >
-                            {this.turtleCustomizationService.getAllTurtles().map(t => {
-                              return (
-                                <option key={t.id} value={t.id}>
-                                  {t.getName()}
-                                </option>
-                              );
-                            })}
-                          </select>
-                        </div>
-                      </div>
-                      <div className="control">
-                        <div className="select">
-                          <select
-                            id="turtleSelector"
-                            value={this.state.userSettings.turtleSize}
-                            onChange={async event => {
-                              const value = parseInt(event.target.value, 10);
-                              await this.userSettingsService.update({ turtleSize: value });
-                              await this.loadData();
-                              this.setRandomCode();
-                            }}
-                          >
-                            <option value={20}>{_T("Extra Small")}</option>
-                            <option value={32}>{_T("Small")}</option>
-                            <option value={40}>{_T("Medium")}</option>
-                            <option value={52}>{_T("Large")}</option>
-                            <option value={72}>{_T("Huge")}</option>
-                          </select>
-                        </div>
-                      </div>
-                    </div>
-
-                    <label className="label">{_T("Personal library")}</label>
-                    <p className="help">
-                      <span>
-                        {_T("You have %d program", {
-                          plural: "You have %d programs",
-                          value: this.state.programCount
-                        })}
-                      </span>
-                    </p>
-                    <div className="field is-grouped is-grouped-multiline">
-                      <p className="control">
-                        <a className="button" onClick={this.doExport}>
-                          {_T("Export")}
-                        </a>
-                      </p>
-                      <p className="control">
-                        <FileSelectorComponent
-                          className={cn({ "is-loading": this.state.isImportingInProgress })}
-                          buttonText={_T("Import")}
-                          onFileTextContentReady={this.onImport}
-                        />
-                      </p>
-                    </div>
-                  </div>
-                </div>
-                <div className="tile is-parent">
-                  <div className="tile is-child box">
+                  <div className="column">
                     {[
                       <LogoExecutorComponent
                         key={`${JSON.stringify(this.state.userSettings)}`} //this is a hack to force component to be created each render in order to not handle prop change event
@@ -322,8 +361,8 @@ export class UserProfilePageComponent extends React.Component<IComponentProps, I
                     ]}
                   </div>
                 </div>
-              </div>
-            )}
+              )}
+          </div>
         </div>
       </div>
     );
